+++ Systems/Physics/MeshDebug.cpp
@@ -120,7 +120,7 @@ void MeshDebug::DrawAllTriangleVoronoiRegions()
 
   if (!mMesh) return;
 
-  uint triCount = mMesh->GetTriangleCount();
+  uint triCount = (uint)mMesh->GetTriangleCount();
 
   for(uint i = 0; i < triCount; ++i)
     DrawTriangleVoronoiRegion(i);
