+++ Systems/Physics/PhysicsMesh.cpp
@@ -122,7 +122,7 @@ void PhysicsMesh::GenerateTree()
     
     // Create the broad phase data.
     BaseBroadPhaseData<uint> data;
-    data.mClientData = triIndex;
+    data.mClientData = (uint)triIndex;
     data.mAabb = ToAabb(tri);
 
     // Insert it into the tree
